backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.1.2
scikit-learn==1.5.2
sentence-transformers==3.0.1
torch>=2.2.0
pdfminer.six==20231228
python-docx==1.1.2
python-multipart==0.0.9
